% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{ATEd}
\alias{ATEd}
\title{Difference in means estimate of the ATE}
\usage{
ATEd(data)
}
\arguments{
\item{data}{data frame containing cols y (positive outcome with zeros) and z (treatment).}
}
\value{
difference in means estimate of the ATE
}
\description{
Computes the ATE by taking the difference in the means.
}
\details{
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
t.test() from base R can be used as an alternative.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
ATEd(sim$data)

}
