% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{ATEo}
\alias{ATEo}
\title{Oracle model estimate of the ATE}
\usage{
ATEo(data)
}
\arguments{
\item{data}{data frame containing cols y (positive outcome with zeros) and z (treatment).}
}
\value{
oracle estimate of ATE
}
\description{
Computes the ATE under the assumption that all strata information are known and true.
}
\details{
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
This serves as a benchmark result under the ideal scenario where all strata are known, and can used to compare with other analysis results.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
ATEo(data)

}
