% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{bounds_ls}
\alias{bounds_ls}
\title{Bounds}
\usage{
bounds_ls(data)
}
\arguments{
\item{data}{data frame containing cols y (positive outcome with zeros) and z (treatment).}
}
\value{
The maximum and lowest possible values for segment proportions, mean, and variance.
}
\description{
Computes parameter bounds to be used in L-BFGS-B mle optimization.
}
\details{
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
Each strata must have at least three observations.\cr
According to RDocumentation, the L-BFGS-B method for optimization is taken from Byrd et. al. (1995), allowing box constraints with
upper and lower bound.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
bounds_ls(sim$data)

}
