% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{hes_ll_ls}
\alias{hes_ll_ls}
\title{Hessian Matrix}
\usage{
hes_ll_ls(par, data, trans = FALSE)
}
\arguments{
\item{par}{vector c(piA, piB, muA1, muA0, muB1, sigma), c(piA, piB/(1-piA), muA1, muA0, muB1, sigma) if trans=TRUE.}

\item{data}{data frame containing columns y (positive outcome with zeros) and z (treatment).}

\item{trans}{boolean signifying if piB has been transformed.}
}
\value{
Hessian matrix for the latent stratification model.
}
\description{
Computes the second order partial derivative with respect to each of the \emph{par} variables, resulting in a Hessian matrix.
}
\details{
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
Sometimes piB is transformed to relative proportions from absolute proportions. This transformation allows the reparameterization
of the piA and piB to allow constraint bounds between 0 and 1 in the optimization procedure. \cr
The returned Hessian is the second order derivative with respect to \eqn{\theta} where
\eqn{\theta} is in the order of piA, piB, muA1, muA0, muB1, and sigma.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
hes_ll_ls(sim$par, sim$data)

# if the strata proportions are in relative sizes
hes_ll_ls(sim$par, sim$data, trans=TRUE)

}
