% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{ls_vcv}
\alias{ls_vcv}
\title{Variance-Covariance Matrix}
\usage{
ls_vcv(par, data, method)
}
\arguments{
\item{par}{vector c(piA, piB, muA1, muA0, muB1, sigma).}

\item{data}{data frame containing cols y (positive outcome with zeros) and z (treatment).}

\item{method}{method used for computation: score, hessian, robust, and bootstrap.}
}
\value{
the variance-covariance matrix based on the specificed method.
}
\description{
Variance-Covariance Matrix
}
\details{
Computes the variance-covariance matrix: \cr
if method = "hessian" then the standard errors are computed by the numeric hessian \cr
if method = "score" then standard errors are computed from the gradient \cr
if method = "robust" then white robust standard errors are computed \cr
if method = "bootstrap" then the standard errors are computed by bootstrap \cr
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
ls_vcv(sim$par, sim$data, "hessian")

}
