% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{start_ls}
\alias{start_ls}
\title{Starting Values}
\usage{
start_ls(data, rand = FALSE)
}
\arguments{
\item{data}{data frame containing cols y (positive outcome with zeros) and z (treatment).}

\item{rand}{boolean controlling whether random adjusted starting mean values are produced.}
}
\value{
starting values used for maximum-likelihood optimization.
}
\description{
Computes the starting values for proportion, mean, and variance for optimization.
The proportions \emph{pi} are transformed to relative sizes.
}
\details{
For the input data frame, column z is the dummy variable for treatment. If z = 1, then the observation has received treatment.
If z = 0, then the observation has not received treatment. \cr
By allowing rand=TRUE, the optim() in mle_ls() can be started in multiple places, providing
random adj values while remaining within reasonable bounds.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
start_ls(sim$data)

# if wish to have multiple random starting values
start_ls(sim$data, rand=TRUE)

}
