% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_functions.R
\name{varATEldelta}
\alias{varATEldelta}
\title{Standard Error of ATE}
\usage{
varATEldelta(par, vcv)
}
\arguments{
\item{par}{vector c(piA, piB, muA1, muA0, muB1, sigma).}

\item{vcv}{variance-covariance matrix of the parameters, can be calculated using ls_vcv().}
}
\value{
standard error of the average treatment effect.
}
\description{
Estimates the standard error and the exponential of the standard error of ATE via the delta method.
}
\details{
The delta method estimates the variance by expanding the function of a random variable through Taylor approximation,
which can be expanded to vectorized calculations.
For a more detailed explanation, see https://www.stata.com/support/faqs/statistics/delta-method/.
}
\examples{
sim = sim_latent_strat(n=10000, piA=0.2, piB=0.1, muA1=5, muA0=4.5, muB1=3, sigma=0.3)
varcovar = ls_vcv(sim$par, sim$data, "hessian")
varATEldelta(sim$par, varcovar)

}
